index {rts}	R Documentation

Extracting and replacing the index of raster time series

Description

index is a generic function for extracting the index of a raster time series (Raster*TS) object and replacing it.

index(x) <- value, can be used to replace index with value, a vector of the same length as the number of raster layers in Raster*TS object .

Usage

## S3 method for class 'RasterStackBrickTS'
index(x, ...)

Arguments

x
A RasterStack or RasterBrick object

...
further arguments passed to methods

Author(s)

Babak Naimi

naimi.b@gmail.com

http://r-gis.net

See Also

index.xts

Examples

file <- system.file("external/ndvi", package="rts")

ndvi <- rts(file) # read the ndvi time series from the specified file

index(ndvi)

[Package rts version 1.0-12 Index]