RasterStackBrickTS-class {rts}	R Documentation

rts classes

Description

A raster time series contains a collection of RasterLayer objects, each corresponds to a time/date.

RasterStackTS and RasterBrickTS classes are created by putting together a RasterStack or RasterBrick object, from the raster package, and an xts object, from the xts package. A RasterStack and RasterBrick represents a collection of RasterLayer objects with the same extent and resolution. An xts object extends the S3 class zoo from the package of the same name. This object provides the index values that is unique and ordered, and also is a time-based class. Currently acceptable classes include: 'Date', 'POSIXct', 'timeDate', as well as 'yearmon' and 'yearqtr' where the index values remain unique.

rts is a subclass of xts class.

Slots

Slots for Raster*TS object:

raster:

object of class RasterStack or RasterBrick

rime:

object of class xts

See also Raster-class for slots in raster.

Author(s)

Babak Naimi

naimi.b@gmail.com

http://r-gis.net

Examples

showClass("RasterStackTS")

[Package rts version 1.0-12 Index]