cellFromXY {rts}	R Documentation

Get cell number from row, column or XY

Description

Get cell number(s) of a Raster*TS object from row and column numbers or X and Y coordinates.

Usage

cellFromRowCol(object, rownr, colnr)
cellFromXY(object, xy)

Arguments

object
a raster time series (Raster*TS) object created by rts

colnr
column number; or vector of column numbers

rownr
row number; or vector of row numbers

xy
matrix of x and y coordinates, or a SpatialPoints or SpatialPointsDataFrame object

Details

These functions are essentially a wrapper to cellFromRowCol and cellFromXY in raster package, work with Raster*TS objects.

Value

row, column or cell number(s). cellFromLine and cellFromPolygon return a list.

Author(s)

Babak Naimi naimi.b@gmail.com http://r-gis.net

Examples

file <- system.file("external/ndvi", package="rts")

rt <- rts(file) # read the ndvi time series from the specified file
cellFromRowCol(rt,15,20)

cellFromRowCol(rt,c(16:20),c(11:15))

cellFromXY(rt,c(645000,57345000))

[Package rts version 1.0-12 Index]