read.rts {rts}	R Documentation

Read raster time Series data from a file

Description

Read a raster time series object from a file.

Usage

read.rts(filename)

Arguments

filename
Filename (character)

Details

This function reads a raster time series object which has been writen by write.rts. Instead of read.rts, the rts function can be used (usage: rts(filename)).

Value

RasterBrickTS

Author(s)

Babak Naimi

naimi.b@gmail.com

http://r-gis.net

See Also

write.rts, rts

Examples

file <- system.file("external/ndvi", package="rts")

rt <- read.rts(file)

rt

# or alternatively:
rt <- rts(file)

[Package rts version 1.0-12 Index]