subset by index {rts}	R Documentation

Subset layers in a raster time series object by index

Description

Extract layers from a Raster*TS object by index (i.e. double bracket, [[).

Details

This function can be used to extract a raster layer or a set of raster lasyers based on the time-index using date-like string. The format must left-specied with respect to the standard ISO:8601 time format "CCYY-MM-DD HH:MM:SS". It is also possible to specify a range of times via the index-based subsetting, using ISO-recommended "/" as the range operator. The basic form is "from/to", where both are optional. If either side is missing, it is interpretted as a request to retrieve raster layers from the beginning, or through the end of the raster time series object. Both subset function and '[[' operator do the same thing.

Value

RasterStackTS or RasterBrickTS.

Methods

x[[i, ...]]

Arguments

i - indicates the layers (represented as a vector of numeric or character string relevant to time index, or by a time-based object).

... - same as ... in subset function in package raster

Author(s)

Babak Naimi

naimi.b@gmail.com

http://r-gis.net

See Also

subset

Examples

file <- system.file("external/ndvi", package="rts")

ndvi <- rts(file) # read the ndvi time series from the specified file

s1 <- ndvi[["2000-01-01/2000-05-31"]]

s1

plot(s1)

[Package rts version 1.0-12 Index]